makefile replacement + corrected some part 1 functions

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: ifridrik <ifridrik@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 17:13:52 by ifridrik          #+#    #+#             */
-/*   Updated: 2023/02/17 14:50:40 by ifridrik         ###   ########.fr       */
+/*   Updated: 2023/03/10 14:08:10 by ifridrik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ char	*ft_strchr(const char *s, int c)
 	int	i;
 
 	i = 0;
-	while (i <= ft_strlen(s))
+	while (i <= (int) ft_strlen(s))
 	{
 		if (s[i] == c)
 			return ((char *) &s[i]);
