+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: ifridrik <ifridrik@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 17:14:13 by ifridrik          #+#    #+#             */
-/*   Updated: 2023/02/17 14:54:41 by ifridrik         ###   ########.fr       */
+/*   Updated: 2023/03/10 14:07:56 by ifridrik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strrchr(const char *s, int c)
 
 	i = 0;
 	last = -1;
-	while (i <= ft_strlen(s))
+	while (i <= (int) ft_strlen(s))
 	{
 		if (s[i] == c)
 			last = i;
